Tests are ok
There is still the usize-u64 ballad for SNPs coords. u64 is ideally better for 32bit architectures...but since I use them as indices for Vec it's completely pointless. Willl need to switch to usize for everything.

+++ src/lib.rs
@@ -281,7 +281,7 @@ mod tests {
                         Mutation { id: "2".to_owned(), pos: csnp3, sequence_ref: vec!(), sequence_alt: vec!(), genotypes : vec!((true, true)), is_indel : false, indel_len: 0});
         let ref mut buffer = VecDeque::<Mutation>::from_iter(muts.into_iter());
         let n_samples = 4;
-        let mut groups : Vec<u32> =  vec![0; n_samples*2];
+        let mut groups : Vec<usize> =  vec![0; n_samples*2];
         indel::IndelRider::count_groups(buffer, 2, &mut groups, n_samples);
         assert_eq!(groups, vec![3, 2, 0, 0, 2, 0, 0, 2]);
     }
@@ -310,7 +310,7 @@ mod tests {
         // The lst indel should be skipped (it is the first Mutation outside our window in this setup.
         let ref mut buffer = VecDeque::<Mutation>::from_iter(muts.into_iter());
         let n_samples = 4;
-        let mut groups : Vec<u32> =  vec![0; n_samples*2];
+        let mut groups : Vec<usize> =  vec![0; n_samples*2];
         indel::IndelRider::count_groups(buffer, 5, &mut groups, n_samples);
         assert_eq!(groups, vec![6, 7, 4, 5, 2, 3, 0, 1]); // NAY
     }
